PDF_Loader/requirements.txt
langchain
openai
flask
python-dotenv
chromadb
pypdf
jinja2 
tiktoken
langchain-community